device/requirements.txt
paho-mqtt>=2.0.0
python-nmap>=0.7.1
requests>=2.31.0
pydantic>=2.5.0
jinja2>=3.1.2
weasyprint>=60.1
markdown>=3.5.1
llama-cpp-python>=0.2.20
Pillow>=10.1.0
RPi.GPIO>=0.7.1
cryptography>=41.0.7
python-dotenv>=1.0.0

